+++ CHANGELOG.md
@@ -2,5 +2,7 @@
 
 * Define configs in modal Extensions.
 * Connect Discord via RPC.
-* Change status microphone (mute, unmute, toggle).
+* Change status microphone (enable, disable, toggle).
+* Change status Deaf (enable, disable, toggle)
 * Documentation from installing
+*